requirements.txt
anyio==3.6.2
click
colorama
fastapi
h11
idna
pydantic
sniffio
starlette
typing_extensions
uvicorn
confuse
mysql-connector-python
